requirements.txt
gradio
torch
diffusers
transformers
python-dotenv